+++ custom_components/pico_rest/const.py
@@ -15,7 +15,6 @@ PLATFORMS = [
     "number",
     "select",
     "time",
-    "light",
     "button",
 ]